Privacy for a couple of structs.
